Finalize Recharge Page
 - Moved Home Realm to front page - It's not relevant to Recharges
 - 1 PP Bonus benefit from E. Infused now separate checkbox
 - Now have 5 recharge boxes
 - Ad/Disadvantages relevant to recharges now under main title + new checkbox
 - QR Code Generation now on Recharge Page
 - Renamed Character Class to Expression on front page
 - Session 1, 2, 3 now Day 1, 2, 3 - Better matches Checkin Process + Session is usually referring to whole weekend, not a day
