+++ api/ExpressedRealms.Characters.Reports/CRB/CrbPages/RechargePage.cs
@@ -2,6 +2,8 @@ using ExpressedRealms.Characters.Reports.CRB.Data;
 using ExpressedRealms.Characters.Reports.CRB.Data.SupportingData;
 using PdfSharp.Drawing;
 using PdfSharp.Pdf;
+using PdfSharp.Pdf.AcroForms;
+using QRCoder;
 
 namespace ExpressedRealms.Characters.Reports.CRB.CrbPages;
 
@@ -19,64 +21,137 @@ internal static class RechargePage
         TextPrintUtilities.PrintStatInfo(
             page,
             powerType,
-            XUnitPt.FromInch(0.47),
+            XUnitPt.FromInch(0.51),
             XUnitPt.FromInch(9.00)
         );
 
         TextPrintUtilities.PrintStatInfo(
             page,
             recharges.ToString(),
-            XUnitPt.FromInch(0.47),
-            XUnitPt.FromInch(7.79)
+            XUnitPt.FromInch(0.51),
+            XUnitPt.FromInch(7.82)
         );
         TextPrintUtilities.PrintStatInfo(
             page,
             pool.ToString(),
-            XUnitPt.FromInch(0.47),
-            XUnitPt.FromInch(6.96)
+            XUnitPt.FromInch(0.51),
+            XUnitPt.FromInch(6.15)
         );
 
         TextPrintUtilities.PrintStatInfo(
             page,
             reportData.WealthInfo.WealthLevel.ToString(),
             XUnitPt.FromInch(3.02),
-            XUnitPt.FromInch(5.88)
+            XUnitPt.FromInch(5.38)
         );
         TextPrintUtilities.PrintStatInfo(
             page,
             reportData.WealthInfo.WealthIncome.ToString("C0"),
             XUnitPt.FromInch(3.02),
-            XUnitPt.FromInch(5.01)
+            XUnitPt.FromInch(4.68)
         );
 
         // Day 1
         if (reportData.BasicInfo.CurrentDay > 1)
         {
-            GenerateRechargeBoxes(page, 0, XUnitPt.FromInch(0.8), XUnitPt.FromInch(9.92));
-            GenerateRechargePoolBoxes(page, 0, XUnitPt.FromInch(0.8), XUnitPt.FromInch(7.32));
+            GenerateRechargeBoxes(page, 0, XUnitPt.FromInch(0.975), XUnitPt.FromInch(9.92));
+            GenerateRechargePoolBoxes(page, 0, XUnitPt.FromInch(0.975), XUnitPt.FromInch(6.505));
         }
         else
         {
-            GenerateRechargeBoxes(page, recharges, XUnitPt.FromInch(0.8), XUnitPt.FromInch(9.92));
-            GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(0.8), XUnitPt.FromInch(7.32));
+            GenerateRechargeBoxes(page, recharges, XUnitPt.FromInch(0.975), XUnitPt.FromInch(9.92));
+            GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(0.975), XUnitPt.FromInch(6.505));
         }
 
         // Day 2
         if (reportData.BasicInfo.CurrentDay > 2)
         {
-            GenerateRechargeBoxes(page, 0, XUnitPt.FromInch(1.65), XUnitPt.FromInch(9.92));
-            GenerateRechargePoolBoxes(page, 0, XUnitPt.FromInch(1.65), XUnitPt.FromInch(7.32));
+            GenerateRechargeBoxes(page, 0, XUnitPt.FromInch(1.74), XUnitPt.FromInch(9.92));
+            GenerateRechargePoolBoxes(page, 0, XUnitPt.FromInch(1.74), XUnitPt.FromInch(6.505));
         }
         else
         {
-            GenerateRechargeBoxes(page, recharges, XUnitPt.FromInch(1.65), XUnitPt.FromInch(9.92));
-            GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(1.65), XUnitPt.FromInch(7.32));
+            GenerateRechargeBoxes(page, recharges, XUnitPt.FromInch(1.74), XUnitPt.FromInch(9.92));
+            GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(1.74), XUnitPt.FromInch(6.505));
         }
 
         GenerateRechargeBoxes(page, recharges, XUnitPt.FromInch(2.51), XUnitPt.FromInch(9.92));
-        GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(2.51), XUnitPt.FromInch(7.32));
+        GenerateRechargePoolBoxes(page, pool, XUnitPt.FromInch(2.51), XUnitPt.FromInch(6.505));
+        
+        DrawQrCode(reportData.BasicInfo.LookupId, document, page);
     }
 
+    
+    private static void DrawQrCode(
+        string content,
+        PdfDocument document,
+        PdfPage page
+    )
+    {
+        
+        var anchor = document.AcroForm.Fields["QrAnchor"] as PdfTextField;
+
+        if (anchor == null)
+            throw new InvalidOperationException("QrAnchor field not found.");
+
+        PdfDictionary? widget;
+
+        // Try /Kids first
+        var kids = anchor.Elements["/Kids"] as PdfArray;
+
+        if (kids != null && kids.Elements.Count > 0)
+        {
+            widget = kids.Elements[0] as PdfDictionary;
+        }
+        else
+        {
+            // Field itself is the widget
+            widget = anchor;
+        }
+
+        if (widget == null)
+            throw new InvalidOperationException("QrAnchor field not found.");
+
+        var rect = widget.Elements.GetRectangle("/Rect");
+
+        var x = XUnitPt.FromPoint(rect.X1);
+        var y = XUnitPt.FromPoint(page.Height.Point - rect.Y2);
+        var width = XUnitPt.FromPoint(rect.Width);
+        var height = XUnitPt.FromPoint(rect.Height);
+
+        using var gfx = XGraphics.FromPdfPage(page);
+
+        
+        using var generator = new QRCodeGenerator();
+        using var data = generator.CreateQrCode(content, QRCodeGenerator.ECCLevel.H);
+
+        var matrix = data.ModuleMatrix;
+        var modules = matrix.Count;
+
+        double moduleSize = XUnitPt.FromPoint(Math.Min(width, height) / modules);
+
+        // Center inside bounds
+        var offsetX = x + (width - modules * moduleSize) / 2;
+        var offsetY = y + (height - modules * moduleSize) / 2;
+
+        for (int row = 0; row < modules; row++)
+        {
+            for (int col = 0; col < modules; col++)
+            {
+                if (matrix[row][col])
+                {
+                    gfx.DrawRectangle(
+                        XBrushes.Black,
+                        XUnitPt.FromPoint(offsetX + col * moduleSize),
+                        XUnitPt.FromPoint(offsetY + row * moduleSize),
+                        XUnitPt.FromPoint(moduleSize),
+                        XUnitPt.FromPoint(moduleSize)
+                    );
+                }
+            }
+        }
+    }
+    
     private static int GetRechargesAndPool(
         ReportData reportData,
         int characterLevel,
@@ -107,8 +182,8 @@ internal static class RechargePage
             TextPrintUtilities.PrintStatInfo(
                 page,
                 "x",
-                XUnitPt.FromInch(0.953),
-                XUnitPt.FromInch(5)
+                XUnitPt.FromInch(0.77),
+                XUnitPt.FromInch(8.30)
             );
             if (energeticallyInfused.Cost == "4pt")
             {
@@ -116,6 +191,13 @@ internal static class RechargePage
             }
             if (energeticallyInfused.Cost == "8pt")
             {
+                // 1 PP Bonus
+                TextPrintUtilities.PrintStatInfo(
+                    page,
+                    "x",
+                    XUnitPt.FromInch(0.77),
+                    XUnitPt.FromInch(7.57)
+                );
                 recharges += 1;
                 pool += characterLevel;
             }
@@ -127,8 +209,8 @@ internal static class RechargePage
             TextPrintUtilities.PrintStatInfo(
                 page,
                 "x",
-                XUnitPt.FromInch(0.473),
-                XUnitPt.FromInch(5)
+                XUnitPt.FromInch(0.77),
+                XUnitPt.FromInch(10.50)
             );
         }
 
@@ -138,14 +220,14 @@ internal static class RechargePage
             TextPrintUtilities.PrintStatInfo(
                 page,
                 "x",
-                XUnitPt.FromInch(0.633),
-                XUnitPt.FromInch(5)
+                XUnitPt.FromInch(0.77),
+                XUnitPt.FromInch(9.775)
             );
             TextPrintUtilities.CrossStampInfo(
                 page,
                 "☥",
-                XUnitPt.FromInch(1.80),
-                XUnitPt.FromInch(4.71)
+                XUnitPt.FromInch(2.155),
+                XUnitPt.FromInch(4.98)
             );
         }
 
@@ -155,8 +237,8 @@ internal static class RechargePage
             TextPrintUtilities.PrintStatInfo(
                 page,
                 "x",
-                XUnitPt.FromInch(0.793),
-                XUnitPt.FromInch(5)
+                XUnitPt.FromInch(0.77),
+                XUnitPt.FromInch(9.045)
             );
             if (weakenedConnection.Cost == "4pt")
             {
@@ -212,7 +294,7 @@ internal static class RechargePage
     )
     {
         var boxCount = 1;
-        for (int y = 0; y < 4; y++)
+        for (int y = 0; y < 5; y++)
         {
             var coordinateX = startX;
             var coordinateY = startY - (y * XUnitPt.FromInch(0.78));
